Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -4,7 +4,7 @@
 int main() {
     float r;
     scanf("%f",&r);
-    int are= (3.14*(r*r));
+    float are= (3.14*(r*r));
     printf("Area: %f",are);
 
     return 0;
